Allow classes to be added to the class whitelist
Previously, the only way to allow elements such as textarea, input, and select
to fire key events was to whitelist them by giving them the 'moustrap' class.
This meant that in order to use mousetrap with existing applications a change
to the html markup might be required.

In order to make this more flexible, and so easier to adopt in existing
applications, we add whitelist() and unwhitelist() methods that can modify
the whitelisted classes easily.

Signed-off-by: Andrew Ardill <andrew.ardill@gmail.com>