Refer to Mousetrap through window in plugins for better compiler support
If Mousetrap and a plugin are compiled together using Closure compiler with the
ADVANCED_OPTIMIZATIONS flag, the compiler needs to be certain that the Mousetrap
object referred to by the plugin is the one created in the main Mousetrap
definition. However, plugin definitions used to refer to Mousetrap without going
through window, so the compiler thought it was a different object. Fix this by
referring to window.Mousetrap also in plugins.