Update mousetrap.js
If there was a space+c binding for instance, it would fire when c is pressed and not consider space. Now, any key will check to ensure previous keys are still held before firing. If window loses focus, it will clear the arrays that hold the held and released keys.