Add an option to determine how the caps lock key is handled
The current default behavior of the caps lock is to automatically
ignore it. This PR adds a configuration option to remove this limiting
behavior.

The fix is to add an options blob to the constructor (more things could
be added in the future as needed). And then this is consumed inside
_characterFromEvent.